clean up engine and fuel code to fully use RF

+++ api/buildcraft/api/fuels/IFuel.java
@@ -15,5 +15,5 @@ public interface IFuel {
 
 	int getTotalBurningTime();
 
-	float getPowerPerCycle();
+	int getPowerPerCycle();
 }
